Modelos
SRFar cambiado campos para que coincida con front

+++ backend/src/models/subjectiverefractionfar.js
@@ -27,11 +27,11 @@ module.exports = (sequelize, DataTypes) => {
             allowNull: true
         },
         axisLE: {
-            type: DataTypes.FLOAT,
+            type: DataTypes.INTEGER,
             allowNull: true
         },
         axisRE: {
-            type: DataTypes.FLOAT,
+            type: DataTypes.INTEGER,
             allowNull: true
         },
         vareachedLE: {
@@ -43,7 +43,7 @@ module.exports = (sequelize, DataTypes) => {
             allowNull: true
         },
         pupilarDistance: {
-            type: DataTypes.FLOAT,
+            type: DataTypes.INTEGER,
             allowNull: false
         }
     }, {
